New FormatEscapeExtension replacing FormatEscaper
Ties together all of the EarlyEscapeNodes and the registered handlers
