+++ src/editor/contents.js
@@ -10,7 +10,6 @@ import { $createLinkNode, $toggleLink } from "@lexical/link"
 import { dispatch, parseHtml } from "../helpers/html_helper"
 import { $isListNode, ListItemNode } from "@lexical/list"
 import { $getNearestNodeOfType } from "@lexical/utils"
-import FormatEscaper from "./contents/format_escaper"
 import Uploader from "./contents/uploader"
 import { $isActionTextAttachmentNode } from "../nodes/action_text_attachment_node"
 
@@ -19,7 +18,6 @@ export default class Contents {
     this.editorElement = editorElement
     this.editor = editorElement.editor
 
-    new FormatEscaper(editorElement).monitor()
   }
 
   insertHtml(html, { tag } = {}) {
